resolved local issues run app, renamed js file to jsx and corrected misplaced semicolons
