+++ capstone-team-1/src/App.jsx
@@ -1,35 +1,37 @@
-import { useEffect, useState } from 'react'
-import './App.css'
-import Footer from './components/Footer/Footer'
-import Header from './components/Header/Header'
+import { useEffect, useState } from 'react';
+import './App.css';
+import Footer from './components/Footer/Footer';
+import Header from './components/Header/Header';
 // import JobList from './components/ProjectList/ProjectList'
 // import data_jobs from './data/joblist.json'
-import BusinessList from './components/BusinessList/BusinessList'
-import businessList from './data/businessList.json'
+import BusinessList from './components/BusinessList/BusinessList';
+import businessList from './data/businessList.json';
+import Register from './Register/Register';
 
 function App() {
-  const [businesses, setBusinesses] = useState([])
-  const [filters, setFilters] = useState([])
+  const [businesses, setBusinesses] = useState([]);
+  const [filters, setFilters] = useState([]);
 
-  useEffect(() => setBusinesses(businessList))
+  useEffect(() => setBusinesses(businessList));
 
   // filter function
   const filterFunc = ({ role, level, tools, languages }) => {
     if (filters.length === 0) {
-      return true
+      return true;
     }
 
-    const tags = [role, level]
+    const tags = [role, level];
 
-    return tags.some((tag) => filters.includes(tag))
-  }
+    return tags.some((tag) => filters.includes(tag));
+  };
 
-  const filteredBusinesses = businesses.filter(filterFunc)
+  const filteredBusinesses = businesses.filter(filterFunc);
 
   return (
-    <div className='w-full min-h-screen'>
+    <div className="w-full min-h-screen">
       <Header />
-      <div className='max-w-7xl mx-auto'>{/* use filter function here */}</div>
+      <Register />
+      <div className="max-w-7xl mx-auto">{/* use filter function here */}</div>
 
       {/* joblist component here */}
       <BusinessList filteredBusinesses={filteredBusinesses} />
@@ -37,7 +39,7 @@ function App() {
       {/* Footer component here */}
       <Footer />
     </div>
-  )
+  );
 }
 
-export default App
+export default App;
